requirements.txt
python-telegram-bot==20.0
requests==2.28.1
fastapi==0.88.0
